+++ packages/pacer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @tanstack/pacer
 
+## 0.18.0
+
+### Minor Changes
+
+- feat: add maxWait option to AsyncRetryer ([#147](https://github.com/TanStack/pacer/pull/147))
+
+## 0.17.3
+
+### Patch Changes
+
+- fix: remove structuredClone from async utils ([`2bb7e70`](https://github.com/TanStack/pacer/commit/2bb7e7096715fbaf4584fe392810dc0fec72f844))
+
+## 0.17.2
+
+### Patch Changes
+
+- Fixed getAbortSignal binding on all utils ([#132](https://github.com/TanStack/pacer/pull/132))
+  Fixed lastResult overwrite bug in AsyncQueuer
+  Updated TanStack Devtools versions
+
 ## 0.17.1
 
 ### Patch Changes